+++ src/components/pages/MenuBar.jsx
@@ -7,7 +7,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import { type RouteDescription } from 'routes/Routes';
+import { type RouteDescription } from 'routes';
 
 import CodimoLink from './logos/CodimoLink';
 import GithubLink from './logos/GithubLink';
@@ -59,7 +59,12 @@ const MenuBar = ({ routes }: Props) => (
         <LocalLinksContainer>
           {routes.map((route, key) => (
             // eslint-disable-next-line react/no-array-index-key
-            <HeaderLink key={key} to={route.path} title={route.title} />
+            <HeaderLink key={key} title={route.title}>
+              {route.children.map((child, key2) => (
+                // eslint-disable-next-line react/no-array-index-key
+                <HeaderLink key={key2} to={`/${route.game}/${child.path}`} title={child.title} />
+              ))}
+            </HeaderLink>
           ))}
         </LocalLinksContainer>