+++ src/containers/GameViewer.jsx
@@ -13,7 +13,9 @@ import styled from 'styled-components';
 
 import MenuBar from 'components/pages/MenuBar';
 import FrontPage from 'components/FrontPage';
-import { type RouteDescription } from 'routes/Routes';
+import { type RouteDescription } from 'routes';
+
+import gameLoader from './gameLoader';
 
 const GameSection = styled.div`
   height: calc(100% - 55px);
@@ -36,15 +38,15 @@ const GameViewer = ({ routes }: Props) => (
           exact={true}
           component={FrontPage}
         />
-        {routes.map((route, key) => (
+        {routes.map((route) => (route.children.map((child, key) => (
           <Route
             // eslint-disable-next-line react/no-array-index-key
             key={key}
-            path={route.path}
+            path={`/${route.game}/${child.path}`}
             exact={route.exact}
-            component={route.main}
+            component={gameLoader(route.game, child.path)}
           />
-        ))}
+        ))))}
       </GameSection>
     </AllHeightBritannia>
   </Router>